+++ src/Iveoles/Image/routes.php
@@ -2,5 +2,5 @@
 
 $route = rtrim(\Config::get('image::route'), '/');
 Route::get($route, function() {
-	App::make('kevbaldwyn.image')->serve();
+	App::make('iveoles.image')->serve();
 });
